blur: Calculate vertices in vertex shader instead of using VBO
Using a VBO for a simple full-screen triangle is unnecessarily
complicated, and because blur rendering is memory-bound, every little
bit of memory bandwidth counts. This removes the VBO in favor of
calculating the UV and clip space coordinates for a full-screen triangle
in the vertex shader, which only takes a few ALU instructions in a
shader that is only run 3 times per rendering step.

When tested with 2 layers of another 6-pass blur implementation, this
saves ~50 µs of rendering time on an Adreno 640 GPU at 1440x3040
resolution.

Change-Id: I738d1a2eccfeb25d66573c3a29416108cbb9b6eb